fix(core): :bug: layoutViewer related changes
layoutViewer related changes

ref: #109
