+++ app/components.registry.js
@@ -1,4 +1,5 @@
 import Components from "./components/component-docs/Components";
+import Layouts from "./components/layout-docs/Layouts";
 import BlankLayoutPage from "./components/layout-docs/page/BlankLayoutPage";
 import CenteredBlankLayoutPage from "./components/layout-docs/page/CenteredBlankLayoutPage";
 import ComplexLayoutPage from "./components/layout-docs/page/ComplexLayoutPage";
@@ -15,8 +16,7 @@ import RightDrawerLayoutPage from "./components/layout-docs/page/RightDrawerLayo
 import ThreeColumnLayoutPage from "./components/layout-docs/page/ThreeColumnLayoutPage";
 import TwoColumnLayoutPage from "./components/layout-docs/page/TwoColumnLayoutPage";
 import VCenteredBlankLayoutPage from "./components/layout-docs/page/VCenteredBlankLayoutPage";
-import SampleLayouts from "./components/layout-docs/SampleLayouts";
-import StyleUtilities from "./components/styles-docs/StyleUtilities";
+import Styles from "./components/styles-docs/Styles";
 
 export const ComponentsRegistry = {
   BlankLayoutPage            : { comp: BlankLayoutPage },
@@ -30,11 +30,11 @@ export const ComponentsRegistry = {
   HCenteredBlankLayoutPage   : { comp: HCenteredBlankLayoutPage },
   HeaderFooterLayoutPage     : { comp: HeaderFooterLayoutPage },
   HeaderLayoutPage           : { comp: HeaderLayoutPage },
+  Layouts                    : { comp: Layouts },
   LeftDrawerLayoutPage       : { comp: LeftDrawerLayoutPage },
   LeftRightDrawerPage        : { comp: LeftRightDrawerPage },
   RightDrawerLayoutPage      : { comp: RightDrawerLayoutPage },
-  SampleLayouts              : { comp: SampleLayouts },
-  StyleUtilities             : { comp: StyleUtilities },
+  Styles                     : { comp: Styles },
   ThreeColumnLayoutPage      : { comp: ThreeColumnLayoutPage },
   TwoColumnLayoutPage        : { comp: TwoColumnLayoutPage },
   VCenteredBlankLayoutPage   : { comp: VCenteredBlankLayoutPage },
